Attempt using big integers - but this does not perform

+++ src/main/java/nl/mvdr/adventofcode/adventofcode2022/day11/Item.java
@@ -1,11 +1,13 @@
 package nl.mvdr.adventofcode.adventofcode2022.day11;
 
+import java.math.BigInteger;
+
 /**
  * Representation of an item which has been stolen by a monkey.
  *
  * @author Martijn van de Rijdt
  */
-record Item(int worryLevel) {
+record Item(BigInteger worryLevel) {
     
     @Override
     public String toString() {
